Queue specification and implementation. Cosmetic changes in Stack implementation. Depth First Search technique implementation. DFS invokes a callback as it visits vertices in its execution. The callback should be supplied by the library user.

+++ stack/value-stack/stack.c
@@ -27,7 +27,7 @@ uint32_t stack_empty(Stack stk)
     return stk.top == -1;
 }
 
-Stack stack_push(Stack  stk, float data, StackResult *result)
+Stack stack_push(Stack  stk, int32_t data, StackResult *result)
 {
     if (stk.top + 1 < stk.size) {
         stk.data[++stk.top] = data;
